+++ packages/server/src/modules/charges/helpers/merge-charges.hepler.ts
@@ -3,7 +3,7 @@ import { Injector } from 'graphql-modules';
 import { BusinessTripEmployeePaymentsProvider } from '@modules/business-trips/providers/business-trips-employee-payments.provider.js';
 import { DocumentsProvider } from '@modules/documents/providers/documents.provider.js';
 import { LedgerProvider } from '@modules/ledger/providers/ledger.provider.js';
-import { TransactionsNewProvider } from '@modules/transactions/providers/transactions-new.provider.js';
+import { TransactionsProvider } from '@modules/transactions/providers/transactions.provider.js';
 import { deleteCharges } from './delete-charges.helper.js';
 
 export const mergeChargesExecutor = async (
@@ -23,7 +23,7 @@ export const mergeChargesExecutor = async (
 
       // update linked transactions
       const replaceTransactionsChargeIdPromise = injector
-        .get(TransactionsNewProvider)
+        .get(TransactionsProvider)
         .replaceTransactionsChargeId({
           replaceChargeID: id,
           assertChargeID: baseChargeID,
