+++ packages/server/src/modules/charts/resolvers/charts.resolver.ts
@@ -1,7 +1,7 @@
 import { GraphQLError } from 'graphql';
 import { ExchangeProvider } from '@modules/exchange-rates/providers/exchange.provider.js';
 import { getTransactionDebitDate } from '@modules/transactions/helpers/debit-date.helper.js';
-import { TransactionsNewProvider } from '@modules/transactions/providers/transactions-new.provider.js';
+import { TransactionsProvider } from '@modules/transactions/providers/transactions.provider.js';
 import { Currency } from '@shared/enums';
 import { dateToTimelessDateString, formatFinancialAmount } from '@shared/helpers';
 import { TimelessDateString } from '@shared/types';
@@ -15,7 +15,7 @@ export const chartsResolvers: ChartsModule.Resolvers = {
       { injector, adminContext: { defaultAdminBusinessId, defaultCryptoConversionFiatCurrency } },
     ) => {
       try {
-        const transactions = await injector.get(TransactionsNewProvider).getTransactionsByFilters({
+        const transactions = await injector.get(TransactionsProvider).getTransactionsByFilters({
           fromDebitDate: filters.fromDate,
           toDebitDate: filters.toDate,
           ownerIDs: [defaultAdminBusinessId],
