+++ packages/server/src/modules/ledger/helpers/dividend-ledger.helper.ts
@@ -1,4 +1,4 @@
-import type { IGetTransactionsByChargeIdsResult } from '@modules/transactions/__generated__/transactions-new.types.js';
+import type { IGetTransactionsByChargeIdsResult } from '@modules/transactions/types.js';
 import { splitFeeTransactions } from './fee-transactions.js';
 
 export function splitDividendTransactions(
