+++ packages/server/src/modules/ledger/resolvers/ledger-generation/common-ledger-generation.resolver.ts
@@ -12,7 +12,7 @@ import {
 import { handleCrossYearLedgerEntries } from '@modules/ledger/helpers/cross-year-ledger.helper.js';
 import { validateExchangeRate } from '@modules/ledger/helpers/exchange-ledger.helper.js';
 import { generateMiscExpensesLedger } from '@modules/ledger/helpers/misc-expenses-ledger.helper.js';
-import { TransactionsNewProvider } from '@modules/transactions/providers/transactions-new.provider.js';
+import { TransactionsProvider } from '@modules/transactions/providers/transactions.provider.js';
 import type { currency } from '@modules/transactions/types.js';
 import type {
   Currency,
@@ -107,7 +107,7 @@ export const generateLedgerRecordsForCommonCharge: ResolverFn<
       : Promise.resolve([]);
 
     const transactionsPromise = gotTransactions
-      ? injector.get(TransactionsNewProvider).transactionsByChargeIDLoader.load(chargeId)
+      ? injector.get(TransactionsProvider).transactionsByChargeIDLoader.load(chargeId)
       : Promise.resolve([]);
 
     const unbalancedBusinessesPromise = injector
