+++ packages/server/src/modules/ledger/resolvers/ledger-generation/conversion-ledger-generation.resolver.ts
@@ -1,6 +1,6 @@
 import { ExchangeProvider } from '@modules/exchange-rates/providers/exchange.provider.js';
 import { storeInitialGeneratedRecords } from '@modules/ledger/helpers/ledgrer-storage.helper.js';
-import { TransactionsNewProvider } from '@modules/transactions/providers/transactions-new.provider.js';
+import { TransactionsProvider } from '@modules/transactions/providers/transactions.provider.js';
 import {
   Currency,
   Maybe,
@@ -56,7 +56,7 @@ export const generateLedgerRecordsForConversion: ResolverFn<
 
     // Get all transactions
     const transactions = await injector
-      .get(TransactionsNewProvider)
+      .get(TransactionsProvider)
       .transactionsByChargeIDLoader.load(chargeId);
     const { mainTransactions, feeTransactions } = splitFeeTransactions(transactions);
