+++ packages/server/src/modules/ledger/resolvers/ledger-generation/dividend-ledger-generation.resolver.ts
@@ -4,7 +4,7 @@ import { ledgerEntryFromBalanceCancellation } from '@modules/ledger/helpers/comm
 import { storeInitialGeneratedRecords } from '@modules/ledger/helpers/ledgrer-storage.helper.js';
 import { generateMiscExpensesLedger } from '@modules/ledger/helpers/misc-expenses-ledger.helper.js';
 import { BalanceCancellationProvider } from '@modules/ledger/providers/balance-cancellation.provider.js';
-import { TransactionsNewProvider } from '@modules/transactions/providers/transactions-new.provider.js';
+import { TransactionsProvider } from '@modules/transactions/providers/transactions.provider.js';
 import { DIVIDEND_WITHHOLDING_TAX_PERCENTAGE } from '@shared/constants';
 import { Maybe, ResolverFn, ResolversParentTypes, ResolversTypes } from '@shared/gql-types';
 import type { LedgerProto, StrictLedgerProto } from '@shared/types';
@@ -63,7 +63,7 @@ export const generateLedgerRecordsForDividend: ResolverFn<
 
     // Get all transactions
     const transactionsPromise = injector
-      .get(TransactionsNewProvider)
+      .get(TransactionsProvider)
       .transactionsByChargeIDLoader.load(chargeId);
 
     const chargeBallanceCancellationsPromise = injector
