+++ packages/server/src/modules/transactions/index.ts
@@ -4,7 +4,7 @@ import transactions from './typeDefs/transactions.graphql.js';
 import { createModule } from 'graphql-modules';
 import { BankDepositTransactionsProvider } from './providers/bank-deposit-transactions.provider.js';
 import { CreditCardTransactionsProvider } from './providers/creditcard-transactions.provider.js';
-import { TransactionsNewProvider } from './providers/transactions-new.provider.js';
+import { TransactionsProvider } from './providers/transactions.provider.js';
 import { CreditCardTransactionsResolvers } from './resolvers/creditcard-transactions.resolver.js';
 import { transactionSuggestionsResolvers } from './resolvers/transaction-suggestions.resolver.js';
 import { transactionsResolvers } from './resolvers/transactions.resolver.js';
@@ -23,7 +23,7 @@ export const transactionsModule = createModule({
   providers: () => [
     BankDepositTransactionsProvider,
     CreditCardTransactionsProvider,
-    TransactionsNewProvider,
+    TransactionsProvider,
   ],
 });
